Rewrite the 'updateOperator' function
The previous version of this function was a mess.
It was necessary to add new codes to this function for
different situations, but it was not possible due to
the unnecessary complexity and lack of transparency in its operation.
So, I tried to rewrite it to eliminate the existing problems.