Fix a bug in the 'round' function
In general, to avoid overflowing the calculator display,
the number of digits that the calculator should show
should not be more than ten, which includes both integer and
decimal parts. But there was a flaw in the function of
the 'round' function, which caused the number of decimal digits
whose integer part is equal to ten or more to not be limited.