Fix wrong calculations in some situations
If the result of a calculation was an integer and the number of digits
of this number was more than ten, in order for this number to fit
in the calculator display and not cause the display to overflow,
instead of displaying the number exponentially, only ten digits
were separated from it. And it was displayed on the screen and
the rest of the digits were thrown away, which caused an error in
the calculation result. Worse than that, if the result of
that calculation was a decimal number, separating only ten digits
from the number would cause the decimal part to be ignored and
thus convert the decimal number into an integer. Therefore,
by converting the number with more than ten digits into
an exponential number, the problem of wrong calculation was solved.

In addition to fixing the problem mentioned above, some minor problems
in the way of displaying numbers (operants) or when division by zero
has happened were also fixed:

- If the user immediately pressed the decimal key for
  the first number (the first operand) and then selected one of
  the operators, that number would be displayed as '0.' on
  the primary and secondary displays. This problem was solved by
  changing the number to '0'.
- It was determined that if the second number is exactly equal to '0',
  the result of the calculation in the division operation is to display
  the message 'Division by 0' to the user. But due to the problem
  mentioned before, the result of division by '0.' was to display
  the message 'Infinity' to the user. Therefore, to solve this problem,
  it was decided that the second number should first be converted
  from a string to a number, so that if it is equal to zero in
  both cases, only the message 'Division by 0'
  will be displayed to the user.

In addition to the above, some unnecessary parts of the code
were merged into other functions or deleted for further simplification.